feat(openchallenges): set non-basic properties when creating a challenge (SMR-327, SMR-372) (#3369)

+++ apps/openchallenges/challenge-service/src/main/java/org/sagebionetworks/openchallenges/challenge/service/model/repository/CustomChallengeRepositoryImpl.java
@@ -169,7 +169,7 @@ public class CustomChallengeRepositoryImpl implements CustomChallengeRepository 
   ) {
     return pf
       .bool(b -> {
-        for (Long operation : query.getOperations()) {
+        for (Integer operation : query.getOperations()) {
           b.should(pf.match().field("operation.id").matching(operation));
         }
       })
@@ -203,7 +203,7 @@ public class CustomChallengeRepositoryImpl implements CustomChallengeRepository 
   ) {
     return pf
       .bool(b -> {
-        for (Long edamConceptId : query.getInputDataTypes()) {
+        for (Integer edamConceptId : query.getInputDataTypes()) {
           b.should(pf.match().field("input_data_types.id").matching(edamConceptId));
         }
       })
